+++ src/main/java/com/verdantartifice/primalmagick/client/events/ClientModLifecycleEvents.java
@@ -28,6 +28,7 @@ import com.verdantartifice.primalmagick.client.renderers.tile.RitualBellTER;
 import com.verdantartifice.primalmagick.client.renderers.tile.RitualLecternTER;
 import com.verdantartifice.primalmagick.client.renderers.tile.RunescribingAltarTER;
 import com.verdantartifice.primalmagick.client.renderers.tile.SanguineCrucibleTER;
+import com.verdantartifice.primalmagick.client.renderers.tile.SpellcraftingAltarTER;
 import com.verdantartifice.primalmagick.client.renderers.tile.WandChargerTER;
 import com.verdantartifice.primalmagick.client.tooltips.ClientAffinityTooltipComponent;
 import com.verdantartifice.primalmagick.common.affinities.AffinityTooltipComponent;
@@ -114,6 +115,7 @@ public class ClientModLifecycleEvents {
         BlockEntityRenderers.register(TileEntityTypesPM.RUNESCRIBING_ALTAR.get(), RunescribingAltarTER::new);
         BlockEntityRenderers.register(TileEntityTypesPM.SANGUINE_CRUCIBLE.get(), SanguineCrucibleTER::new);
         BlockEntityRenderers.register(TileEntityTypesPM.AUTO_CHARGER.get(), AutoChargerTER::new);
+        BlockEntityRenderers.register(TileEntityTypesPM.SPELLCRAFTING_ALTAR.get(), SpellcraftingAltarTER::new);
     }
     
     private static void registerItemProperties(FMLClientSetupEvent event) {