+++ src/main/java/com/verdantartifice/primalmagick/client/events/EntityRendererEvents.java
@@ -25,6 +25,7 @@ import com.verdantartifice.primalmagick.client.renderers.entity.model.PixieModel
 import com.verdantartifice.primalmagick.client.renderers.entity.model.SpellMineModel;
 import com.verdantartifice.primalmagick.client.renderers.entity.model.SpellProjectileModel;
 import com.verdantartifice.primalmagick.client.renderers.models.ModelLayersPM;
+import com.verdantartifice.primalmagick.client.renderers.tile.model.SpellcraftingAltarRingModel;
 import com.verdantartifice.primalmagick.common.entities.EntityTypesPM;
 
 import net.minecraft.client.model.HumanoidModel;
@@ -103,5 +104,6 @@ public class EntityRendererEvents {
         event.registerLayerDefinition(ModelLayersPM.SPELL_MINE, SpellMineModel::createBodyLayer);
         event.registerLayerDefinition(ModelLayersPM.SPELL_PROJECTILE, SpellProjectileModel::createBodyLayer);
         event.registerLayerDefinition(ModelLayersPM.TREEFOLK, () -> LayerDefinition.create(HumanoidModel.createMesh(CubeDeformation.NONE, 0.0F), 64, 64));
+        event.registerLayerDefinition(ModelLayersPM.SPELLCRAFTING_ALTAR_RING, SpellcraftingAltarRingModel::createBodyLayer);
     }
 }