+++ src/main/java/com/verdantartifice/primalmagick/client/fx/particles/ParticleTypesPM.java
@@ -51,4 +51,8 @@ public class ParticleTypesPM {
             return NoteEmitterParticleData.CODEC;
         }
     });
+    public static final RegistryObject<SimpleParticleType> SPELLCRAFTING_RUNE_U = PARTICLE_TYPES.register("spellcrafting_rune_u", () -> new SimpleParticleType(true));
+    public static final RegistryObject<SimpleParticleType> SPELLCRAFTING_RUNE_V = PARTICLE_TYPES.register("spellcrafting_rune_v", () -> new SimpleParticleType(true));
+    public static final RegistryObject<SimpleParticleType> SPELLCRAFTING_RUNE_T = PARTICLE_TYPES.register("spellcrafting_rune_t", () -> new SimpleParticleType(true));
+    public static final RegistryObject<SimpleParticleType> SPELLCRAFTING_RUNE_D = PARTICLE_TYPES.register("spellcrafting_rune_d", () -> new SimpleParticleType(true));
 }